+++ CHANGES.rst
@@ -2,8 +2,44 @@ Changelog
 =========
 
 
-1.0a3 (unreleased)
+1.0a4 (unreleased)
 ------------------
+
+- Nothing changed yet.
+
+
+1.0a3 (2019-11-28)
+------------------
+
+- Update french translations
+  [mpeeters]
+
+- Added feature : force resync a meeting
+  [odelaere]
+
+- factorize sync methods
+  [odelaere]
+
+- Drop Plone 5.1 and 5.0 support
+  [mpeeters]
+
+- Managed info_points_formatting_tal in sync
+  [odelaere]
+
+- Manage last modification date sync format for meeting
+  [jjaumotte]
+
+- Add `collective.fingerpointing` to the package dependencies
+  [mpeeters]
+
+- Add tests for utils functions
+  [mpeeters]
+
+- Publish demo profile content
+  [mpeeters]
+
+- Add tests for faceted criteria
+  [mpeeters]
 
 - Add tests for utils, item and institution views
   [mpeeters]
@@ -28,6 +64,7 @@ Changelog
 
 - Redirect on faceted view after importing a meeting
   [mpeeters]
+
 - They may be 0 or 1 or more Representatives in charge of an item
   [odelaere]